requirements.txt
requests
openstacksdk
Flask
google-generativeai